Support Threads.Condition variables in Select.
Note that you need to unlock the condition inside the body of your
select, which is super annoying and kind of makes it not-worth-it to
support Condition variables.

It would probably be better to just have an assertion error if you try
to wait on a condition variable...

BUT, anyway, this commit fixes the waiting to support condition
variables correctly, I'm pretty sure! :)